+++ src/types/locatorApi.ts
@@ -109,7 +109,7 @@ export interface LocalAuthority {
   recyclingUri: string;
   hwrcUri: string;
   gardenWasteUri: string;
-  bulkyWasteUri: string;
+  bulkyWasteUri?: string;
   bulkyWaste?: BulkyWasteCollection[];
   properties: { [key in PROPERTY_TYPE_EN]?: LocalAuthorityProperty[] };
 }
